djangoapi/packageApp/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2
requests
django-cors-headers